+++ include/settings/armour_sets.h
@@ -11,11 +11,11 @@ class ArmourSets {
 		           ArmourSet tanner_set,
 		           ArmourSet stonecutter_set);
 
-		[[nodiscard]] double smeltingYieldBonus() const;
-		[[nodiscard]] double woodworkingYieldBonus() const;
-		[[nodiscard]] double leatherworkingYieldBonus() const;
-		[[nodiscard]] double weavingYieldBonus() const;
-		[[nodiscard]] double stonecuttingYieldBonus() const;
+		[[nodiscard]] float smeltingYieldBonus() const;
+		[[nodiscard]] float woodworkingYieldBonus() const;
+		[[nodiscard]] float leatherworkingYieldBonus() const;
+		[[nodiscard]] float weavingYieldBonus() const;
+		[[nodiscard]] float stonecuttingYieldBonus() const;
 
 		ArmourSet& getSmelterSet();
 		ArmourSet& getWoodworkerSet();
